requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
# Pour pgvector support
pgvector==0.2.4

# Settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Cloudinary
cloudinary==1.38.0

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3

# Face Recognition (installation optionnelle - voir README)
# face-recognition==1.3.0
# dlib==19.24.2

# OCR
# easyocr==1.7.0  # Commenté temporairement (lourd à installer)

# HTTP Requests
requests==2.31.0

# CORS
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
